Add make and enter page
