+++ public/style.css
@@ -1,3 +1,8 @@
+/* 정렬 */
+.center {
+	text-align: center;
+}
+
 /* 간격 */
 .marginBottom {
 	margin-bottom: 30px;
