+++ src/views/MakePage.vue
@@ -1,199 +1,62 @@
 <template>
-  <div id="main-container" class="container">
-    <div id="join" v-if="!session">
-      <div id="join-dialog" class="jumbotron vertical-center">
-        <h1>Join a video session</h1>
-        <div class="form-group">
-          <p>
-            <label>Participant</label>
-            <input v-model="myUserName" class="form-control" type="text" required />
-          </p>
-          <p>
-            <label>Session</label>
-            <input v-model="mySessionId" class="form-control" type="text" required />
-          </p>
-          <p class="text-center">
-            <button class="btn btn-lg btn-success" @click="joinSession()">
-              Join!
-            </button>
-          </p>
-        </div>
-      </div>
+  <div id="join-dialog" class="jumbotron vertical-center">
+    <div class="largeTitle">MAKE A ROOM</div>
+    <div class="center marginBottom">
+      <p class="middleTitle">참여자</p>
+      <input v-model="UserName" class="form-control" type="text" required />
     </div>
-
-    <div id="session" v-if="session">
-      <div id="session-header">
-        <h1 id="session-title">{{ mySessionId }}</h1>
-        <input class="btn btn-large btn-danger" type="button" id="buttonLeaveSession" @click="leaveSession"
-          value="Leave session" />
-      </div>
-      <div id="main-video" class="col-md-6">
-        <user-video :stream-manager="mainStreamManager" />
-      </div>
-      <div id="video-container" class="col-md-6">
-        <user-video :stream-manager="publisher" @click.native="updateMainVideoStreamManager(publisher)" />
-        <user-video v-for="sub in subscribers" :key="sub.stream.connection.connectionId" :stream-manager="sub"
-          @click.native="updateMainVideoStreamManager(sub)" />
-      </div>
+    <div class="center marginBottom">
+      <p class="middleTitle">방 이름 입력하기</p>
+      <input v-model="RoomName" class="form-control" type="text" required />
+    </div>
+    <div class="center marginBottom">
+          <p class="middleTitle">방 인원 선택하기</p>
+          <button @click="setParticipantCount(1)" :class="{'btn-selected': participantCount === 1, 'btn-empty': participantCount !== 1}">1명</button>
+          <button @click="setParticipantCount(2)" :class="{'btn-selected': participantCount === 2, 'btn-empty': participantCount !== 2}">2명</button>
+          <button @click="setParticipantCount(4)" :class="{'btn-selected': participantCount === 4, 'btn-empty': participantCount !== 4}">4명</button>
+        </div>
+    <div class="center marginBottom">
+      <button class="btn btn-lg btn-success" @click="joinSession">방 만들기</button>
     </div>
   </div>
 </template>
 
 <script>
-import axios from "axios";
-import { OpenVidu } from "openvidu-browser";
-import UserVideo from "@/components/video/UserVideo";
-
-axios.defaults.headers.post["Content-Type"] = "application/json";
-
-// const APPLICATION_SERVER_URL = process.env.NODE_ENV === 'production' ? '' : 'http://localhost:8080/';
-
-const APPLICATION_SERVER_URL = 'https://4cutstudio.store/';
+import { v4 as uuidv4 } from 'uuid';
 
 export default {
-  name: "App",
-
-  components: {
-    UserVideo,
-  },
-
+  name: "MakePage",
   data() {
     return {
-      // OpenVidu objects
-      OV: undefined,
-      session: undefined,
-      mainStreamManager: undefined,
-      publisher: undefined,
-      subscribers: [],
-
-      // Join form
-      mySessionId: "SessionA",
-      myUserName: "Participant" + Math.floor(Math.random() * 100),
+      UserName: "Participant" + Math.floor(Math.random() * 100),
+      RoomName: "",
+      SessionId: "",
+      participantCount: 1, 
     };
   },
-
   methods: {
+    setParticipantCount(count) {
+      this.participantCount = count; 
+    },
     joinSession() {
-      // --- 1) Get an OpenVidu object ---
-      this.OV = new OpenVidu();
-
-      // --- 2) Init a session ---
-      this.session = this.OV.initSession();
-
-      // --- 3) Specify the actions when events take place in the session ---
-
-      // On every new Stream received...
-      this.session.on("streamCreated", ({ stream }) => {
-        const subscriber = this.session.subscribe(stream);
-        this.subscribers.push(subscriber);
-      });
-
-      // On every Stream destroyed...
-      this.session.on("streamDestroyed", ({ stream }) => {
-        const index = this.subscribers.indexOf(stream.streamManager, 0);
-        if (index >= 0) {
-          this.subscribers.splice(index, 1);
-        }
-      });
-
-      // On every asynchronous exception...
-      this.session.on("exception", ({ exception }) => {
-        console.warn(exception);
-      });
-
-      // --- 4) Connect to the session with a valid user token ---
-
-      // Get a token from the OpenVidu deployment
-      this.getToken(this.mySessionId).then((token) => {
-
-        // First param is the token. Second param can be retrieved by every user on event
-        // 'streamCreated' (property Stream.connection.data), and will be appended to DOM as the user's nickname
-        this.session.connect(token, { clientData: this.myUserName })
-          .then(() => {
-
-            // --- 5) Get your own camera stream with the desired properties ---
-
-            // Init a publisher passing undefined as targetElement (we don't want OpenVidu to insert a video
-            // element: we will manage it on our own) and with the desired properties
-            let publisher = this.OV.initPublisher(undefined, {
-              audioSource: undefined, // The source of audio. If undefined default microphone
-              videoSource: undefined, // The source of video. If undefined default webcam
-              publishAudio: true, // Whether you want to start publishing with your audio unmuted or not
-              publishVideo: true, // Whether you want to start publishing with your video enabled or not
-              resolution: "640x480", // The resolution of your video
-              frameRate: 30, // The frame rate of your video
-              insertMode: "APPEND", // How the video is inserted in the target element 'video-container'
-              mirror: false, // Whether to mirror your local video or not
-            });
-
-            // Set the main video in the page to display our webcam and store our Publisher
-            this.mainStreamManager = publisher;
-            this.publisher = publisher;
-
-            // --- 6) Publish your stream ---
-
-            this.session.publish(this.publisher);
-          })
-          .catch((error) => {
-            console.log("There was an error connecting to the session:", error.code, error.message);
-          });
-      });
-
-      window.addEventListener("beforeunload", this.leaveSession);
-    },
-
-    leaveSession() {
-      // --- 7) Leave the session by calling 'disconnect' method over the Session object ---
-      if (this.session) this.session.disconnect();
-
-      // Empty all properties...
-      this.session = undefined;
-      this.mainStreamManager = undefined;
-      this.publisher = undefined;
-      this.subscribers = [];
-      this.OV = undefined;
-
-      // Remove beforeunload listener
-      window.removeEventListener("beforeunload", this.leaveSession);
-    },
-
-    updateMainVideoStreamManager(stream) {
-      if (this.mainStreamManager === stream) return;
-      this.mainStreamManager = stream;
-    },
-
-    /**
-     * --------------------------------------------
-     * GETTING A TOKEN FROM YOUR APPLICATION SERVER
-     * --------------------------------------------
-     * The methods below request the creation of a Session and a Token to
-     * your application server. This keeps your OpenVidu deployment secure.
-     *
-     * In this sample code, there is no user control at all. Anybody could
-     * access your application server endpoints! In a real production
-     * environment, your application server must identify the user to allow
-     * access to the endpoints.
-     *
-     * Visit https://docs.openvidu.io/en/stable/application-server to learn
-     * more about the integration of OpenVidu in your application server.
-     */
-    async getToken(mySessionId) {
-      const sessionId = await this.createSession(mySessionId);
-      return await this.createToken(sessionId);
-    },
-
-    async createSession(sessionId) {
-      const response = await axios.post(APPLICATION_SERVER_URL + 'api/sessions', { customSessionId: sessionId }, {
-        headers: { 'Content-Type': 'application/json', },
-      });
-      return response.data; // The sessionId
-    },
-
-    async createToken(sessionId) {
-      const response = await axios.post(APPLICATION_SERVER_URL + 'api/sessions/' + sessionId + '/connections', {}, {
-        headers: { 'Content-Type': 'application/json', },
-      });
-      return response.data; // The token
+      if (this.RoomName) {
+        this.SessionId = uuidv4();
+        this.$router.push({
+          name: 'Enter',
+          params: {
+            // sessionId: this.SessionId,
+            // userName: this.UserName,
+            // participantCount: this.participantCount, 
+          },
+          query: {
+            sessionId: this.SessionId,
+            userName: this.UserName,
+            participantCount: this.participantCount, 
+          },
+        });
+      } else {
+        alert("방 이름을 입력하세요");
+      }
     },
   },
 };
